updated DB connection to make use of new TickItDB
